% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{add_class}
\alias{add_class}
\title{Add a new class attribute to the input object.}
\usage{
add_class(.obj, .class)
}
\arguments{
\item{.obj}{R object.}

\item{.class}{String with the desired class name.}
}
\value{
Input object with additional class \code{.class}.
}
\description{
Add a new class attribute to the input object.
}
\concept{utility_public}
