% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_distribution}
\alias{check_distribution}
\title{Check that distribution is sums up to one.}
\usage{
check_distribution(
  .data,
  .do.norm = NA,
  .laplace = 1,
  .na.val = 0,
  .warn.zero = F,
  .warn.sum = T
)
}
\arguments{
\item{.data}{Numeric vector of values.}

\item{.do.norm}{One of the three values - NA, T or F. If NA then check for distrubution (sum(.data) == 1)
and normalise if needed with the given laplace correction value. if T then do normalisation and laplace
correction. If F then don't do normalisaton and laplace correction.}

\item{.laplace}{Value for the laplace correction.}

\item{.na.val}{Replace all NAs with this value.}

\item{.warn.zero}{if T then the function checks if in the resulted vector (after normalisation)
are any zeros, and prints a warning message if there are some.}

\item{.warn.sum}{if T then the function checks if the sum of resulted vector (after normalisation)
is equal to one, and prints a warning message if not.}
}
\value{
Numeric vector.
}
\description{
Check if the given .data is a distribution and normalise it if necessary with an optional laplace correction.
}
\concept{utility_public}
