% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{.quant_column_choice}
\alias{.quant_column_choice}
\title{Get a column's name using the input alias}
\usage{
.quant_column_choice(x, .verbose = T)
}
\arguments{
\item{x}{Character vector of length 1.}

\item{.verbose}{If TRUE then print any issues or wanings.}
}
\description{
Get a column's name using the input alias
}
\concept{utility_private}
