% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{filter_barcodes}
\alias{filter_barcodes}
\title{Filter data by barcodes}
\usage{
filter_barcodes(.data, .barcodes)
}
\arguments{
\item{.data}{The data to be processed. Can be \link{data.frame},
\link{data.table}, or a list of these objects.

Every object must have columns in the immunarch compatible format.
\link{immunarch_data_format}

Competent users may provide advanced data representations:
DBI database connections, Apache Spark DataFrame from \link{copy_to} or a list
of these objects. They are supported with the same limitations as basic objects.

Note: each connection must represent a separate repertoire.}

\item{.barcodes}{Character vector. Vector of barcodes to use for
filtering.}
}
\description{
Given the input data and a vector of barcodes,
the function returns the input with cells which has
at least one of the given barcode. Columns with clonotype counts
and proportions are changed accordingly to the filtered barcodes.
}
\examples{
\dontrun{
# Create a vector with barcodes
barcodes = c("AAAGCAACATCACAAC-1", "GGCGACTTCATATCGG-1")
# Filter data and create a new list with datasets
your_data_filt = list()
your_data_filt$data = filter_barcodes(your_data$data, c("AAAGCAACATCACAAC-1", "GGCGACTTCATATCGG-1"))
your_data_filt$meta = your_data$meta
}

}
