% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{has_class}
\alias{has_class}
\title{Check class}
\usage{
has_class(.data, .class)
}
\arguments{
\item{.data}{Any R object.}

\item{.class}{Character vector. Specifies a class name to check against.}
}
\value{
Logical value.
}
\description{
Check class
}
\concept{utility_public}
