% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{matrixdiagcopy}
\alias{matrixdiagcopy}
\title{Copy the upper matrix triangle to the lower one}
\usage{
matrixdiagcopy(.mat)
}
\arguments{
\item{.mat}{Matrix.}
}
\value{
Matrix with its upper tri part copied to the lower tri part.
}
\description{
Copy the upper matrix triangle to the lower one
}
\concept{utility_private}
