% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap_coef}
\alias{overlap_coef}
\alias{jaccard_index}
\alias{tversky_index}
\alias{morisita_index}
\alias{cosine_sim}
\alias{horn_index}
\title{Various measures for overlapping sets of objects.}
\usage{
overlap_coef(.x, .y)

jaccard_index(.x, .y)

tversky_index(.x, .y, .a = .5, .b = .5)

cosine_sim(.x, .y, .quant)

morisita_index(.x, .y, .quant)

horn_index(.x, .y)
}
\arguments{
\item{.x, .y}{Character vectors or data frames, sets of objects to intersect, see "Details"
for more information.}

\item{.a, .b}{Alpha and beta parameters for Tversky Index. Default values gives the Jaccard index measure.}

\item{.quant}{Character. Name of the column with frequencies.}
}
\value{
Matrix with overlap values.
}
\description{
Functions for computing similarity between two vectors or sets.

- Overlap cofficient is a similarity measure related to the Jaccard index that measures the overlap between two sets, and is defined as the size of the intersection divided by the smaller of the size of the two sets.

- Jaccard index is a statistic used for comparing the similarity and diversity of sample sets.

- Tversky index is an asymmetric similarity measure on sets that compares a variant to a prototype and is related to Jaccard index.

- Cosine similarity is a measure of similarity between two vectors of an inner product space that measures the cosine of the angle between them.

- Morisita's overlap index is a statistical measure of dispersion of individuals in a population. It is used to compare overlap among samples (Morisita 1959). This formula is based on the assumption that increasing the size of the samples will increase the diversity because it will include different habitats (i.e. different faunas).

- Horn's overlap index is based on Shannon's entropy.

Use the \link{repOverlap} function for computing similarities of clonesets.
}
\details{
For \code{cosine_sim} .x and .y are two numeric vectors.
For \code{overlap_coef}, \code{jaccard_index} and \code{tversky_index} .x and .y are character vectors.
For \code{morisita_index} and \code{horn_index} .x and .y are two data frames or data tables with character
vectors as first columns and numeric (i.e., number of such elements) as second.
}
\seealso{
\link{repOverlap}
}
