% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{public_matrix}
\alias{public_matrix}
\title{Get a matrix with public clonotype frequencies}
\usage{
public_matrix(.data)
}
\arguments{
\item{.data}{Public repertoire, an output from \link{pubRep}.}
}
\value{
Matrix with per-sample clonotype counts / proportions only.
}
\description{
Get a matrix with public clonotype frequencies
}
