% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{pubRepFilter}
\alias{pubRepFilter}
\alias{publicRepertoireFilter}
\title{Filter out clonotypes from public repertoires}
\usage{
pubRepFilter(.pr, .meta, .by, .min.samples = 1)
}
\arguments{
\item{.pr}{Public repertoires, an output from \link{pubRep}.}

\item{.meta}{Metadata file.}

\item{.by}{Character vector. Names of groups to group by.}

\item{.min.samples}{Integer. Filter out clonotypes with number of samples below than this number.}
}
\value{
Data frame with filtered clonotypes.
}
\description{
Filter out clonotypes from public repertoires
}
