% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_analysis.R
\name{repOverlapAnalysis}
\alias{repOverlapAnalysis}
\title{Analysis of overlaps or any distance matrices}
\usage{
repOverlapAnalysis(
  .data,
  .method = ("hclust"),
  .scale = default_scale_fun,
  .raw = T,
  .perp = 1,
  .theta = 0.1,
  .eps = 0.01,
  .k = 2
)
}
\arguments{
\item{.data}{Any distance matrix between pairs of repertoires. You can also pass your
output from \code{\link{repOverlap}}.}

\item{.method}{A string that defines the type of analysis to perform.}

\item{.scale}{A function to scale the data before passing it to the MDS algorithm.}

\item{.raw}{A logical value. Pass TRUE if you want to receive raw output of clustering
or dimensionality reduction function of choice. Pass FALSE if you want to receive
processed output that can be subjected to visualisation with \code{\link{vis}} function.}

\item{.perp}{A numerical value, t-SNE parameter, see \code{\link{immunr_tsne}}.}

\item{.theta}{A numerical value, t-SNE parameter, see \code{\link{immunr_tsne}}.}

\item{.eps}{A numerical value, DBscan epsylon parameter, see \code{\link{immunr_dbscan}}.}

\item{.k}{The number of clusters to create, passed as \code{k} to \link[factoextra]{hcut} or as \code{centers} to \link{kmeans}.}
}
\value{
Work in progress.
}
\description{
The \code{\link{repOverlapAnalysis}} function contains advanced data
analysis methods. You can use several clustering and dimensionality reduction
techniques in order to investigate further the difference between repertoires
provided.

To cluster a subset of similar data with \code{\link{repOverlapAnalysis}} you can
perform hierarchical clustering, k-means or dbscan ('hclust', 'kmeans', 'dbscan'
respectively).

To reduce dimensions, for example, to select features for subsequent analysis,
you can execute the multidimensional scaling or t-sne algorithms ('mds' and 'tsne'
respectively).
}
