% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{repSave}
\alias{repSave}
\title{Save repertoire files to the disk.}
\usage{
repSave(.data, .path, .format = c("immunarch", "vdjtools"), .compress = T)
}
\arguments{
\item{.data}{An R dataframe, a list of R dataframes or a list with \code{data} and
\code{meta} where first element is a list of dataframes and the latter is a dataframe
with metadata.}

\item{.path}{A string with the path to the output directory. It should include file
name if a single dataframe is provided to .data argument.}

\item{.format}{A string with desirable format specification. Current options are
"immunarch" and "vdjtools".}

\item{.compress}{A boolean value. Defines whether the output will be compressed or not.}
}
\description{
The \code{repSave} function is deigned to save your data to the disk
in desirable format. Currently supports "immunarch" and "vdjtools" file formats.
}
\details{
It is not necessary to create directories beforehand. If the provided directory
does not exist it will be created automatically.
}
