% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{split_immunoseq_files}
\alias{split_immunoseq_files}
\title{Split all-samples-in-one ImmunoSEQ files by sample}
\usage{
split_immunoseq_files(.big.file, .output.folder)
}
\arguments{
\item{.big.file}{Path to the input ImmunoSEQ file with the "sample_name" column with information about receptors' source sample.}

\item{.output.folder}{Path to the folder to write sample repertoires to.}
}
\description{
Adaptive Biotechnologies ImmunoSEQ platform files are often downloaded as a big text file
with receptor data for all samples from the experiment.
Information about the source of each receptor is located in the "sample_name" column of the file.
Such format is impractical for using in R analysis pipelines and require preprocessing
to split the file into smaller files each corresponding to a specific sample.
The \code{split_immunoseq_files} function splits the file into smaller files and writes them in the same ImmunoSEQ
format to the output folder provided by the user.

If you want to just load the big file as separated data frames into R to use with immunarch, just use the \link{repLoad} function
that automatically detectes the ImmunoSEQ format and splits the big file.
}
\examples{
\dontrun{
# Split the file into smaller files
split_immunoseq_files("~/file/path/immunoseq_big.tsv", "~/file/path/experiment1")

# If you just want to load all files as separate data frames to use with immunarch,
# just use the repLoad function
immdata = repLoad("~/file/path/immunoseq_big.tsv")
}
}
\seealso{
\link{repLoad}
}
