% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis.immunr_inc_overlap}
\alias{vis.immunr_inc_overlap}
\title{Visualise incremental overlaps}
\usage{
\method{vis}{immunr_inc_overlap}(.data, .target = 1, .grid = F, .ncol = 2, ...)
}
\arguments{
\item{.data}{Output from the \link{repOverlap} function that uses "top" methods.}

\item{.target}{Index of a repertoire to plot. Omitted if .grid is TRUE.}

\item{.grid}{Logical. If TRUE then plot all similarities in a grid.}

\item{.ncol}{Numeric. Number of columns in the resulting grid.}

\item{...}{Not used here.}
}
\description{
Visualise incremental overlaps
}
\examples{
\dontrun{
data(immdata)
tmp = repOverlap(immdata$data, "inc+overlap")
vis(tmp)
vis(tmp, .grid = T)
}

}
\seealso{
\link{repOverlap}
}
