% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis.immunr_ov_matrix}
\alias{vis.immunr_ov_matrix}
\alias{vis.immunr_gu_matrix}
\title{Overlap / gene usage distance visualisation}
\usage{
\method{vis}{immunr_ov_matrix}(.data, .plot = c("heatmap", "heatmap2", "circos", "radar"), ...)
}
\arguments{
\item{.data}{Output from \link{repOverlap} or \link{geneUsageAnalysis}.}

\item{.plot}{A string specifying the plot type:

- "heatmap" for heatmaps using \link{vis_heatmap};

- "heatmap2" for heatmaps using \link{vis_heatmap2};

- "circos" for circos plots using \link{vis_circos};

- "radar" for polar area plots using \link{vis_radar}.}

\item{...}{Other arguments are passed through to the underlying plotting function:

- "heatmap" - passes arguments to \link{vis_heatmap};

- "heatmap2" - passes arguments to \link{vis_heatmap2} and \link{heatmap} from the "heatmap3" package;

- "circos" - passes arguments to \link{vis_circos} and \link{chordDiagram} from the "circlize" package;

- "radar"- passes arguments to \link{vis_radar}.}
}
\description{
Visualise matrices with overlap values or gene usage distances among samples.
For details see links below.
}
