% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis.immunr_public_repertoire}
\alias{vis.immunr_public_repertoire}
\title{Public repertoire visualisation}
\usage{
\method{vis}{immunr_public_repertoire}(.data, .plot = c("freq", "clonotypes"), ...)
}
\arguments{
\item{.data}{Public repertoire, an output from \link{pubRep}.}

\item{.plot}{A string specifying the plot type:

- "freq" for visualisation of the distribution of occurrences of clonotypes
and their frequencies using \link{vis_public_frequencies}.

- "clonotypes" for visualisation of public clonotype frequenciy correlations between pairs of
samples using \link{vis_public_clonotypes}}

\item{...}{Further arguments passed \link{vis_public_frequencies} or \link{vis_public_clonotypes},
depending on the ".plot" argument.}
}
\description{
Public repertoire visualisation
}
\examples{
\dontrun{
pr = pubRep(immdata$data)
vis(pr, "freq")
vis(pr, "freq", .type="boxplot")
vis(pr, "freq", .type="none")
vis(pr, "freq", .type="mean")
vis(pr, "freq", .by = "Status", .meta=immdata$meta)

vis(pr, "clonotypes", 1, 2)
}
}
