% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis.immunr_public_statistics}
\alias{vis.immunr_public_statistics}
\title{Visualise sharing of clonotypes among samples}
\usage{
\method{vis}{immunr_public_statistics}(.data, ...)
}
\arguments{
\item{.data}{Public repertoire - an output from the \link{pubRep} function.}

\item{...}{Other arguments passsed directly to \link{upset}.}
}
\description{
Visualise public clonotype frequencies.
}
