% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis_heatmap2}
\alias{vis_heatmap2}
\title{Visualisation of matrices using heatmap3-based heatmaps}
\usage{
vis_heatmap2(
  .data,
  .grid.size = 1.5,
  .title = NULL,
  .labs = NULL,
  col = colorRampPalette(c("#67001f", "#d6604d", "#f7f7f7", "#4393c3", "#053061"))(1024),
  ...
)
}
\arguments{
\item{.data}{Input matrix. Column names and row names (if presented) will be used as names for labs.}

\item{.grid.size}{Line width of the grid between cells.}

\item{.title}{The text for the plot's title (same as the "main" argument in \link[heatmap3]{heatmap3}).}

\item{.labs}{A character vector of length two with names for x-axis and y-axis, respectively.}

\item{col}{A vector specifying the colors (same as the "col" argument in \link[heatmap3]{heatmap3}).}

\item{...}{Other arguments for the \link[heatmap3]{heatmap3} function.}
}
\description{
Visualise matrices with the functions based on the \link[heatmap3]{heatmap3}
package with minimum amount of arguments.
}
\seealso{
\link{vis}, \link{repOverlap}.
}
