% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis_treemap}
\alias{vis_treemap}
\title{Visualisation of data frames and matrices using treemaps}
\usage{
vis_treemap(.data, ...)
}
\arguments{
\item{.data}{Input data frame or matrix.}

\item{...}{Other arguments passed to \link{treemap}.}
}
\description{
Visualisation of data frames and matrices using treemaps
}
